Set DefaultShow from os.Stderr to os.Stdout
For commands that use Stdout to print out precisely formated data, such as
CSV files, it may be best to point DefaultShow to Stderr, since in these
cases Show() etal are generally used to print out informational messages, not
the data itself.  However, I am finding these commands the exception, and am
setting the default output for Show() etal to be Stdout.